Fix Markdown crash
Fixes issue https://github.com/diamondburned/dissent/issues/275.

The fix was to copy almost all of yuin/goldmark's list code over to the
repository just because they didn't expose one very simple option to not
allow '+' be parsed.

The solution is extremely fragile and may break in the future.
We should not update goldmark.